requirements.txt
streamlit==1.31.0
google-generativeai==0.3.2
requests==2.31.0
python-dotenv==1.0.0
pillow==10.2.0
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0

